Rename src/robothub-depthai to src/robothub-oak
